+++ Liviano.CLI/Program.cs
@@ -60,7 +60,7 @@ namespace Liviano.CLI
         static string hdPath = "m/84'/0'/0'/0/0"; // Default BIP84 / Bitcoin / 1st account / receive / 1st pubkey
         // static string server = "";
         static double amount = 0.00;
-        static int feeSatsPerByte = 1;
+        static decimal feeSatsPerByte = 1m;
         static int accountIndex = -1;
         // static string accountName = null;
         static string walletId = "";
@@ -127,7 +127,7 @@ namespace Liviano.CLI
                 {"pass|passphrase=", "Passphrase", (string v) => passphrase = v},
                 // {"s|server=", "Server", (string v) => server = v},
                 {"amt|amount=", "Amount to send", (string v) => amount = double.Parse(v)},
-                {"fee|fee-sats-per-byte=", "Fees in satoshis per byte", (string v) => feeSatsPerByte = int.Parse(v)},
+                {"fee|fee-sats-per-byte=", "Fees in satoshis per byte", (string v) => feeSatsPerByte = decimal.Parse(v)},
                 {"acci|account-index=", "Account to send from", (string v) => accountIndex = int.Parse(v)},
                 // {"accn|account-name=", "Account to send from", (string v) => accountName = v},
                 {"naccname|new-account-name=", "New account name", (string v) => newAccName = v},
